Fix DMRG convergence, add iPEPS/iDMRG/benchmarks, default float64 (#21)
* Fix DMRG SVD: absorb singular values into MPS tensors

The SVD step in _svd_and_truncate_site returned U and Vh but never
absorbed the singular values back into the MPS, silently dropping
them at every bond update. This corrupted the wavefunction and
caused energy oscillation / non-convergence.

Now singular values are absorbed into B (during L→R sweeps) or A
(during R→L sweeps) to maintain proper canonical form. Also make
trivial environment dtype respect JAX x64 mode instead of
hardcoding float32.

Update the cylinder example to enable JAX_ENABLE_X64 and use
slightly larger bond dimensions for the 2D systems.

Before: all three cylinder configs oscillate and fail to converge.
After:  4x2 converges in 2 sweeps (matches ED to 3e-8),
        6x3 in 5 sweeps, 8x4 in 7 sweeps.


* Update docs with hardware extras and 2D cylinder DMRG example

- README.md: replace bare `pip install` with hardware-specific extras
  (cuda12, cuda13, tpu, metal); add 2D cylinder DMRG example section
- docs/guide/installation.md: expand GPU section with all extras
  including CUDA local variants, TPU, Metal, and ROCm guidance
- docs/guide/algorithms/dmrg.md: add 2D Heisenberg cylinder example


* Add platform markers to CUDA and TPU extras

CUDA and TPU packages only have Linux wheels, so `--all-extras`
fails on macOS. Add `sys_platform == 'linux'` markers to skip
these extras on non-Linux platforms.


* Stop tracking .claude/settings.local.json

Remove the file from git and add it to .gitignore. This is a
user-local Claude Code config that should not be committed.


* Fix macOS CI: skip jax-metal extra to avoid Metal backend errors

jax-metal is experimental and causes UNIMPLEMENTED errors on the
GitHub Actions macOS runner. Use `--extra dev` instead of
`--all-extras` so tests run on CPU without the Metal plugin.


* Add infinite cylinder Heisenberg model via iDMRG super-site MPO

Implement build_bulk_mpo_heisenberg_cylinder() which bundles Ly spins
into a single super-site (d=2^Ly, D_w=3Ly+2), encoding within-ring
bonds as an on-site term and between-ring bonds as NN MPO channels.
Only even Ly is supported since odd circumference frustrates AFM order.

Also change iDMRG dtype defaults from float32 to float64, and fix
within-ring bond double-counting for Ly=2 periodic rings.

Ly=2 energy per spin (-0.5780) matches the 2-leg ladder literature
value to 5 significant figures.


* Fix dtype test to handle JAX without x64 mode

CI runs without JAX_ENABLE_X64, so float64 silently truncates to
float32. Make the test check the expected dtype based on whether
x64 is actually enabled.


* Add 2-site iPEPS simple update with checkerboard CTMRG

A 1x1 unit cell cannot represent 2-sublattice Neel order, so the simple
update gets stuck at the ferromagnetic product state (E~0.5). This adds a
2-site (A/B) checkerboard unit cell with proper alternating-sublattice
CTMRG environment computation and mixed-environment RDMs for energy
evaluation.

New iPEPSConfig.unit_cell="2site" dispatches through ipeps() to the full
pipeline: 2-site simple update → ctm_2site → compute_energy_ctm_2site.
Heisenberg energy now reaches E < -0.3, confirming Neel order is captured.


* Default to float64 throughout and document JAX x64 gotchas

Switch AutoMPO default dtype from float32 to float64, replace
os.environ x64 hacks in examples with jax.config.update, add
gotchas doc page, save benchmarks incrementally, and remove
obsolete PLAN files.


* Add benchmarks/results/ to .gitignore


---------

+++ src/tnjax/algorithms/auto_mpo.py
@@ -325,7 +325,7 @@ def _w_matrices_to_symmetric_mpo(
     d: int,
     phys_charges: np.ndarray,
     bond_charges: list[np.ndarray],
-    dtype: Any = jnp.float32,
+    dtype: Any = jnp.float64,
     name: str = "AutoMPO_sym",
 ) -> TensorNetwork:
     """Wrap W-matrices into a TensorNetwork of SymmetricTensor nodes.
@@ -391,7 +391,7 @@ def _w_matrices_to_symmetric_mpo(
 def _w_matrices_to_mpo(
     w_matrices: list[np.ndarray],
     d: int,
-    dtype: Any = jnp.float32,
+    dtype: Any = jnp.float64,
     name: str = "AutoMPO",
 ) -> TensorNetwork:
     """Wrap a list of W-matrices into a TensorNetwork of DenseTensor nodes.
@@ -561,7 +561,7 @@ class AutoMPO:
         self,
         compress: bool = False,
         compress_tol: float = 1e-12,
-        dtype: Any = jnp.float32,
+        dtype: Any = jnp.float64,
         symmetric: bool = False,
         phys_charges: np.ndarray | None = None,
     ) -> TensorNetwork:
@@ -650,7 +650,7 @@ def build_auto_mpo(
     site_ops: dict[str, np.ndarray] | None = None,
     compress: bool = False,
     compress_tol: float = 1e-12,
-    dtype: Any = jnp.float32,
+    dtype: Any = jnp.float64,
     symmetric: bool = False,
     phys_charges: np.ndarray | None = None,
 ) -> TensorNetwork:
